Add Flags and Blocks
