--- app/src/components/ExploreContainer.tsx
@@ -1,15 +1,0 @@
-import './ExploreContainer.css';
-
-interface ContainerProps {
-  name: string;
-}
-
-const ExploreContainer: React.FC<ContainerProps> = ({ name }) => {
-  return (
-    <div className="container">
-      <strong>{name}</strong>
-    </div>
-  );
-};
-
-export default ExploreContainer;
