+++ app/src/components/assets/Asset.css
@@ -1,0 +1,5 @@
+/* CSS for all Asset pages*/
+
+.imgIcon{
+    margin: 5px;
+}
