+++ app/src/pages/Page.css
@@ -1,0 +1,21 @@
+/* Use this as a global css file 
+ * As long as the content is wrapped in <Page /> elements, this file is included
+*/
+
+.container {
+    margin: 10px;
+}
+
+.row {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    width: 100%;
+}
+
+.column {
+    display: flex;
+    flex-direction: column;
+    flex-basis: 100%;
+    flex: 1;
+}
